Various patches to get the source code build and run today

+++ src/dirent.c
@@ -21,7 +21,7 @@ extern "C"
 
 struct DIR
 {
-    long                handle; /* -1 for failed rewind */
+    intptr_t            handle; /* -1 for failed rewind */
     struct _finddata_t  info;
     struct dirent       result; /* d_name null iff first time */
     char                *name;  /* null-terminated char string */
@@ -42,7 +42,7 @@ DIR *opendir(const char *name)
         {
             strcat(strcpy(dir->name, name), all);
 
-            if((dir->handle = (long) _findfirst(dir->name, &dir->info)) != -1)
+            if((dir->handle = _findfirst(dir->name, &dir->info)) != -1)
             {
                 dir->result.d_name = 0;
             }
